requirements.txt
asgiref==3.8.1
dill==0.3.9
Django==5.1.5
djangorestframework==3.15.2
filelock==3.17.0
fsspec==2025.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
pbr==6.1.0
psutil==6.1.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
qiskit==1.3.2
qiskit-aer==0.16.0
rustworkx==0.16.0
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.1
torch==2.6.0
typing_extensions==4.12.2
tzdata==2025.1